+++ test/runtime.spec.ts
@@ -1,36 +1,50 @@
-import * as oar from '../src/runtime';
 import * as jsc from 'jsverify';
 import * as _ from 'lodash';
 import * as assert from 'assert';
 import { promisify } from 'util';
+import {
+  createMakerWith,
+  Make,
+  makeAllOf,
+  makeAny,
+  makeArray,
+  MakeError,
+  makeNumber,
+  makeObject,
+  makeOneOf,
+  makeOptional,
+  Maker,
+  makeString
+} from '../src/make';
+import { toJSON, ValueClass } from '../src/value-class';
+import { pmap, set } from '../src/runtime';
 
 interface ShapeOfTestClass {
   a: ReadonlyArray<string>;
   b: string;
 }
-class TestClass extends oar.ValueClass<TestClass, ShapeOfTestClass, 1> {
-  static make(v: ShapeOfTestClass): oar.Make<TestClass> {
+class TestClass extends ValueClass<TestClass, ShapeOfTestClass, 1> {
+  static make(v: ShapeOfTestClass): Make<TestClass> {
     return makeTestClass(v);
   }
   public b: string;
   public a: ReadonlyArray<string>;
   constructor(v: ShapeOfTestClass) {
     super();
-    const value = oar
-      .makeObject({
-        a: oar.makeArray(oar.makeString()),
-        b: oar.makeString()
+    const value = makeObject({
+        a: makeArray(makeString()),
+        b: makeString()
       })(v)
       .success();
     Object.assign(this, value);
   }
 }
-const makeTestClass: oar.Maker<ShapeOfTestClass, TestClass> = oar.createMakerWith(TestClass);
+const makeTestClass: Maker<ShapeOfTestClass, TestClass> = createMakerWith(TestClass);
 
 describe('Make', () => {
   describe('success', () => {
     it('uses custom error handler', () => {
-      const value = oar.Make.error([{ path: [], error: 'xxx' }]);
+      const value = Make.error([{ path: [], error: 'xxx' }]);
       expect(value.success(e => 'got error')).toEqual('got error');
     });
   });
@@ -39,7 +53,7 @@ describe('Make', () => {
 describe('pmap', () => {
   jsc.property('leaves object unchanged when no matches', jsc.json, async dict => {
     const clone = _.cloneDeep(dict);
-    const mapped = await oar.pmap(
+    const mapped = await pmap(
       dict,
       (n: any): n is string => false,
       async (n: string) => assert.fail('cant happen')
@@ -51,14 +65,14 @@ describe('pmap', () => {
 
   jsc.property('leaves object unchanged when no changes', jsc.json, async dict => {
     const clone = _.cloneDeep(dict);
-    const mapped = await oar.pmap(dict, (n: any): n is any => true, async (n: any) => n);
+    const mapped = await pmap(dict, (n: any): n is any => true, async (n: any) => n);
     expect(mapped === dict).toBeTruthy();
     expect(mapped).toEqual(clone);
     return true;
   });
 
   jsc.property('maps objects', jsc.dict(jsc.string), async dict => {
-    const mapped = await oar.pmap(
+    const mapped = await pmap(
       dict,
       (n: any): n is string => _.isString(n),
       async (n: string) => n.toUpperCase()
@@ -74,7 +88,7 @@ describe('pmap', () => {
 
   it('sets ValueClasses', async () => {
     const value = TestClass.make({ a: ['a value'], b: 'b value' }).success();
-    const mapped = await oar.pmap(
+    const mapped = await pmap(
       value,
       (n: any): n is string[] => _.isArray(n),
       async (n: string[]) => n.map(w => w.toUpperCase())
@@ -88,18 +102,18 @@ describe('pmap', () => {
   it('throws when setting ValueClass with incorrect value', async () => {
     const value = TestClass.make({ a: ['a value'], b: 'b value' }).success();
     try {
-      await oar.pmap(
+      await pmap(
         value,
         (n: any): n is string[] => _.isArray(n),
         async (n: string[]) => n.map(w => 1 as any)
       );
     } catch (e) {
-      expect(e).toBeInstanceOf(oar.MakeError);
+      expect(e).toBeInstanceOf(MakeError);
     }
   });
 
   jsc.property('maps arrays', jsc.array(jsc.asciistring), async arr => {
-    const mapped = await oar.pmap(
+    const mapped = await pmap(
       arr,
       (n: any): n is string => _.isString(n),
       async (n: string) => n.toUpperCase()
@@ -142,8 +156,7 @@ describe('pmap', () => {
       'does not trigger unhandledRejection for leaked promises',
       jsc.json,
       async json => {
-        await oar
-          .pmap(
+        await pmap(
             json,
             (n: any): n is any => matchesPredicate(n),
             async (n: any) => {
@@ -158,8 +171,7 @@ describe('pmap', () => {
     jsc.property('waits for all promises when all succeed', jsc.json, async json => {
       let promises = 0;
       let done = 0;
-      await oar
-        .pmap(
+      await pmap(
           json,
           (n: any): n is any => matchesPredicate(n),
           async (n: any) => {
@@ -178,29 +190,29 @@ describe('pmap', () => {
 
 describe('makeOneOf', () => {
   it('succeeds if only one match', () => {
-    const fun = oar.makeOneOf(
-      oar.makeObject({ a: oar.makeString() }),
-      oar.makeObject({ a: oar.makeNumber() })
+    const fun = makeOneOf(
+      makeObject({ a: makeString() }),
+      makeObject({ a: makeNumber() })
     );
     expect(fun({ a: 1 }).success()).toEqual({ a: 1 });
   });
 
   it('prefers matching ValueClass', () => {
-    const fun = oar.makeOneOf(oar.makeAny(), TestClass.make);
+    const fun = makeOneOf(makeAny(), TestClass.make);
     const test = TestClass.make({ a: ['a'], b: 'b' }).success();
     expect(fun(test).success()).toEqual(test);
   });
 
   it('fails if multiple values classes match', () => {
-    const fun = oar.makeOneOf(oar.makeAny(), TestClass.make, TestClass.make);
+    const fun = makeOneOf(makeAny(), TestClass.make, TestClass.make);
     const test = TestClass.make({ a: ['a'], b: 'b' }).success();
     expect(fun(test).isSuccess()).toBeFalsy();
   });
 
   it('fails on multiple matches', () => {
-    const fun = oar.makeOneOf(
-      oar.makeObject({ a: oar.makeString() }),
-      oar.makeObject({ a: oar.makeAny() })
+    const fun = makeOneOf(
+      makeObject({ a: makeString() }),
+      makeObject({ a: makeAny() })
     );
     expect(fun({ a: 'x' }).isSuccess()).toBeFalsy();
   });
@@ -208,9 +220,9 @@ describe('makeOneOf', () => {
 
 describe('makeAllOf', () => {
   it('applies all makers to value in succession', () => {
-    const fun = oar.makeAllOf(
-      oar.makeObject({ a: oar.makeString() }, oar.makeAny()),
-      oar.makeObject({}, oar.makeAllOf())
+    const fun = makeAllOf(
+      makeObject({ a: makeString() }, makeAny()),
+      makeObject({}, makeAllOf())
     );
     expect(fun({ a: 'xxx', b: 1 }).success()).toEqual({ a: 'xxx', b: 1 });
   });
@@ -218,7 +230,7 @@ describe('makeAllOf', () => {
 
 describe('makeAny', () => {
   jsc.property('allows anything', jsc.json, async item => {
-    const fun = oar.makeAny();
+    const fun = makeAny();
     expect(fun(item).success()).toEqual(item);
     if (item && typeof item === 'object') {
       expect(fun(item).success() !== item).toBeTruthy();
@@ -229,39 +241,39 @@ describe('makeAny', () => {
 
 describe('makeArray', () => {
   it('keeps the order', () => {
-    const fun = oar.makeArray(oar.makeString());
+    const fun = makeArray(makeString());
     expect(fun(['a', 'b']).success()).toEqual(['a', 'b']);
   });
 });
 
 describe('makeObject', () => {
   it('drops unknown properties if told to', () => {
-    const fun = oar.makeObject({ a: oar.makeNumber() });
+    const fun = makeObject({ a: makeNumber() });
     expect(fun({ a: 1, missing: 'a' }, { unknownField: 'drop' }).success()).toEqual({ a: 1 });
   });
 
   it('disallows extra fields', () => {
-    const fun = oar.makeObject({ a: oar.makeNumber() });
+    const fun = makeObject({ a: makeNumber() });
     expect(fun({ a: 1, missing: 'a' }).isError()).toBeTruthy();
   });
 
   it('allows fields', () => {
-    const fun = oar.makeObject({ a: oar.makeNumber() });
+    const fun = makeObject({ a: makeNumber() });
     expect(fun({ a: 1 }).isSuccess()).toBeTruthy();
   });
 
   it('allows additional props', () => {
-    const fun = oar.makeObject({}, oar.makeString());
+    const fun = makeObject({}, makeString());
     expect(fun({ a: 'xxx' }).success()).toEqual({ a: 'xxx' });
   });
 
   it('prefers specified field to additional props', () => {
-    const fun = oar.makeObject({ a: oar.makeNumber() }, oar.makeString());
+    const fun = makeObject({ a: makeNumber() }, makeString());
     expect(fun({ a: 1 }).success()).toEqual({ a: 1 });
   });
 
   it('allows undefined props', () => {
-    const fun = oar.makeObject({ a: oar.makeOptional(oar.makeNumber()) });
+    const fun = makeObject({ a: makeOptional(makeNumber()) });
     expect(fun({ a: undefined }).success()).toEqual({});
   });
 });
@@ -270,7 +282,7 @@ describe('ValueClass', () => {
   describe('toJSON', () => {
     it('returns a plain javascript object', () => {
       const value = TestClass.make({ b: 'a', a: ['a'] }).success();
-      const json = oar.toJSON<TestClass, ShapeOfTestClass>(value);
+      const json = toJSON<TestClass, ShapeOfTestClass>(value);
       expect(json instanceof TestClass).toBeFalsy();
       expect(json.a).toEqual(['a']);
       expect(json.b).toEqual('a');
@@ -280,7 +292,7 @@ describe('ValueClass', () => {
   describe('set', () => {
     it('creates a new value', () => {
       const value = TestClass.make({ a: ['a'], b: 'original value' }).success();
-      const newValue = oar.set(value, { b: 'new value' }).success();
+      const newValue = set(value, { b: 'new value' }).success();
       expect(newValue.b).toEqual('new value');
       expect(newValue.a).toEqual(['a']);
       expect(value.b).toEqual('original value');
@@ -290,7 +302,7 @@ describe('ValueClass', () => {
 
 describe('createMakerWith', () => {
   function maker() {
-    return oar.createMakerWith<ShapeOfTestClass, TestClass>(TestClass);
+    return createMakerWith<ShapeOfTestClass, TestClass>(TestClass);
   }
 
   it('uses the class instance if given as parameter', () => {
